requirements.txt
torch==1.7.1
torchvision==0.8.2
ml-collections==0.1.0
scikit-learn==0.24.1
imageio
pyyaml
statsmodels
tensorflow
tensorboard
git+https://github.com/kevinzakka/torchkit.git
git+https://github.com/kevinzakka/x-magical.git
